requirements.txt
# Django Framework
Django==5.1.3

# REST API Support
djangorestframework==3.15.2
djangorestframework-simplejwt~=5.3.1  # For JWT authentication

# Celery
celery~=5.4.0
redis~=5.2.0

# Report Generation
WeasyPrint==63.1  # For generating PDFs

# Firebase (For notifications)
firebase-admin~=6.6.0
google-auth==2.36.0

